Update integrations.html (#215)
I replaced pl.com with prayerletters.com in the description for the prayerletters.com integration. In a couple of places for the description of the prayerletters.com integration, we had replaced prayerletters.com with pl.com. pl.com is actually a site in use and we didn't want people navigating there by accident.